feat: Add devcontainer configuration for Freqtrade
Add the devcontainer configuration file for Freqtrade, which includes the necessary settings for running Freqtrade in a development container. This allows developers to easily set up their development environment for Freqtrade.

The devcontainer configuration file specifies the name of the dev container, the Docker image to use, and customizations for Visual Studio Code.

The changes include:
- Adding the devcontainer.json file with the necessary settings.

Closes #123